Init Commit
Add basic functionality for the loot drop generation.

The current loot pool included in this project has not been tested at the time of this commit but is more for testing purposes and will most likely be revised at a later time.

The timer function is also to be tested as the syntax may not be correct.